improved cage solver

+++ solver/src/rules.rs
@@ -18,7 +18,7 @@ impl Default for Rules {
 pub struct Cages {
     pub cages: Vec<u32>,
     #[serde(with = "a81")]
-    pub cells: [u32; 81],
+    pub cells: [usize; 81],
 }
 
 impl Default for Cages {
